// apps/admin/src/lib/api/generated/models/connectorPaginationMode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.51
 */

export type ConnectorPaginationMode = typeof ConnectorPaginationMode[keyof typeof ConnectorPaginationMode];


export const ConnectorPaginationMode = {
  none: 'none',
  cursor: 'cursor',
  page: 'page',
  offset: 'offset',
  next_url: 'next_url',
} as const;