// apps/admin/src/lib/api/generated/models/connectorAuthMode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.51
 */

export type ConnectorAuthMode = typeof ConnectorAuthMode[keyof typeof ConnectorAuthMode];


export const ConnectorAuthMode = {
  none: 'none',
  bearer: 'bearer',
  api_key: 'api_key',
} as const;